+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning], with the exception that minor rel
 - Refactored state preparation circuit synthesis code to utilize the new `PureFaultSet` and `CNOTCircuit` classes. ([#443]) ([**@pehamtom**])
 - Refactored encoding circuit synthesis code to utilize the new `PureFaultSet` and `CNOTCircuit` classes. ([#443]) ([**@pehamtom**])
 - Renamed `StatePrepCircuit` class to `FaultyStatePrepCircuit`, reflecting its new role in combining circuit and fault information. ([#443]) ([**@pehamtom**])
+- Added `NoiseModel` class for applying noise to a given stim circuit. ([#453]) ([**@pehamtom**])
 
 ## [1.9.0] - 2025-03-14
 
@@ -32,6 +33,7 @@ _📚 Refer to the [GitHub Release Notes](https://github.com/munich-quantum-tool
 <!-- PR links -->
 
 [#443]: https://github.com/munich-quantum-toolkit/qecc/pull/443
+[#453]: https://github.com/munich-quantum-toolkit/qecc/pull/453
 
 <!-- Contributor -->